% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{billbydate}
\alias{billbydate}
\title{按日期同步}
\usage{
billbydate(fdate)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{应收单}

\item{fnumber}{S202307040026}
}
\value{
返回值
}
\description{
按日期同步
}
\examples{
billbydate("2023-07-11")
}
