% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{billbyno}
\alias{billbyno}
\title{按单同步单据}
\usage{
billbyno(token, category, fnumber)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{应收单}

\item{fnumber}{S202307040026}
}
\value{
返回值
}
\description{
按单同步单据
}
\examples{
billbyno("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","应收单","S202307040026")
}
