% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{billdateupdate}
\alias{billdateupdate}
\title{修改单据状态}
\usage{
billdateupdate(token, category, fnumber)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{应收单}

\item{fnumber}{S202307040026}
}
\value{
返回值
}
\description{
修改单据状态
}
\examples{
billdateupdate("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","应收单","S202307040026")
}
