% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsbill.R
\name{ecsbill_sync2}
\alias{ecsbill_sync2}
\title{输入日期同步所有单据}
\usage{
ecsbill_sync2(startDate, endDate)
}
\arguments{
\item{startDate}{开始时间}

\item{endDate}{结束时间}
}
\value{

}
\description{
输入日期同步所有单据
}
\examples{
ecsbill_sync2(2023-07-27 00:00:00,2023-07-28 09:34:24)
}
