% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{erpbillquery}
\alias{erpbillquery}
\title{erp单据查询}
\usage{
erpbillquery(token, category, fnumber)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{其他出库单}

\item{fnumber}{D202305310016}
}
\value{
返回值
}
\description{
erp单据查询
}
\examples{
erpbillquery("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","其他出库单","D202305310016")
}
