% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{logquery}
\alias{logquery}
\title{日志查询}
\usage{
logquery(token, fnumber)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{fnumber}{D202305310016}
}
\value{
返回值
}
\description{
日志查询
}
\examples{
logquery("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","D202305310016")
}
