% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticeShipment.R
\name{noticeShipmentByDate_sync}
\alias{noticeShipmentByDate_sync}
\title{按日期同步}
\usage{
noticeShipmentByDate_sync(Ftoken, FDate, FName)
}
\arguments{
\item{Ftoken}{中台数据库token}

\item{FDate}{业务日期}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按日期同步
}
\examples{
noticeShipmentByDate_sync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','2023-07-05','赛普集团新账套')
}
