% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticeShipment.R
\name{noticeShipmentByNumber_sync}
\alias{noticeShipmentByNumber_sync}
\title{按单同步}
\usage{
noticeShipmentByNumber_sync(Ftoken, FNumber, FName)
}
\arguments{
\item{Ftoken}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单同步
}
\examples{
noticeShipmentByNumber_sync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','D202306200021','赛普集团新账套')
}
