% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticeShipment.R
\name{noticeShipmentLog_query}
\alias{noticeShipmentLog_query}
\title{日志查询}
\usage{
noticeShipmentLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志查询
}
\examples{
noticeShipmentLog_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,D202306200021)
}
