% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherInStock.R
\name{otherInStockStatus_upload}
\alias{otherInStockStatus_upload}
\title{修改订单状态}
\usage{
otherInStockStatus_upload(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
修改订单状态
}
\examples{
otherInStockStatus_upload('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','PI202307240003','赛普集团新账套')
}
