% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseOrder.R
\name{purchaseOrderByNumber}
\alias{purchaseOrderByNumber}
\title{按单查询}
\usage{
purchaseOrderByNumber(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
按单查询
}
\examples{
purchaseOrderByNumber(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,SP202307030002)
}
