% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseOrder.R
\name{purchaseOrderLog_query}
\alias{purchaseOrderLog_query}
\title{日志查询}
\usage{
purchaseOrderLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志查询
}
\examples{
purchaseOrderLog_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,SP202307030002)
}
