% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseOrder.R
\name{purchaseOrderStatus_upload}
\alias{purchaseOrderStatus_upload}
\title{修改订单状态}
\usage{
purchaseOrderStatus_upload(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
修改订单状态
}
\examples{
purchaseOrderStatus_upload(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,SP202307030002,赛普集团新账套)
}
