% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchasesBilling.R
\name{purchasesBillingByDate_query}
\alias{purchasesBillingByDate_query}
\title{按日期查询}
\usage{
purchasesBillingByDate_query(token, FStartDate)
}
\arguments{
\item{token}{中台数据库token}

\item{FStartDate}{业务日期}
}
\value{

}
\description{
按日期查询
}
\examples{
purchasesBillingByDate_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,2023-07-05)
}
