% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchasesBilling.R
\name{purchasesBillingByNumber_query}
\alias{purchasesBillingByNumber_query}
\title{按单查询}
\usage{
purchasesBillingByNumber_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
按单查询
}
\examples{
purchasesBillingByNumber_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','SD202307060001')
}
