% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseStorage.R
\name{purchaseStorageByNumber_sync}
\alias{purchaseStorageByNumber_sync}
\title{按单同步}
\usage{
purchaseStorageByNumber_sync(Ftoken, FNumber, FName)
}
\arguments{
\item{Ftoken}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单同步
}
\examples{
purchaseStorageByNumber_sync(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,PI202307030001,赛普集团新账套)
}
