% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnNotice.R
\name{returnNoticeLog_query}
\alias{returnNoticeLog_query}
\title{日志查询}
\usage{
returnNoticeLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志查询
}
\examples{
returnNoticeLog_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,R202304110001)
}
