% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPurchase.R
\name{returnPurchaseByNumber_sync}
\alias{returnPurchaseByNumber_sync}
\title{按单同步}
\usage{
returnPurchaseByNumber_sync(Ftoken, FNumber, FName)
}
\arguments{
\item{Ftoken}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单同步
}
\examples{
returnPurchaseByNumber_sync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','SR202307250005','赛普集团新账套')
}
