% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRequest.R
\name{returnRequestByDate_sync}
\alias{returnRequestByDate_sync}
\title{按日期同步}
\usage{
returnRequestByDate_sync(Ftoken, FDate, FName)
}
\arguments{
\item{Ftoken}{中台数据库token}

\item{FDate}{业务日期}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按日期同步
}
\examples{
returnRequestByDate_sync(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,2023-07-05,赛普集团新账套)
}
