% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnRequest.R
\name{returnRequestByNumber_query}
\alias{returnRequestByNumber_query}
\title{按单查询}
\usage{
returnRequestByNumber_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
按单查询
}
\examples{
returnRequestByNumber_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','SR202307250005')
}
