% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{saleOrderLog_query}
\alias{saleOrderLog_query}
\title{日志查询}
\usage{
saleOrderLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志查询
}
\examples{
saleOrderLog_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,202307030001)
}
