% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{saleOrderStatus_upload}
\alias{saleOrderStatus_upload}
\title{修改订单状态}
\usage{
saleOrderStatus_upload(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
修改订单状态
}
\examples{
saleOrderStatus_upload(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,202307030001,赛普集团新账套)
}
