% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOut.R
\name{saleOutErpDataByFNumber_query}
\alias{saleOutErpDataByFNumber_query}
\title{ERP数据查询}
\usage{
saleOutErpDataByFNumber_query(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
ERP数据查询
}
\examples{
saleOutErpDataByFNumber_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,D202306200021,赛普集团新账套)
}
