% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesBilling.R
\name{salesBillingErpDataByFNumber_query}
\alias{salesBillingErpDataByFNumber_query}
\title{ERP数据查询}
\usage{
salesBillingErpDataByFNumber_query(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
ERP数据查询
}
\examples{
salesBillingErpDataByFNumber_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','S202307200003','赛普集团新账套')
}
