% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{srcbilldatequery}
\alias{srcbilldatequery}
\title{源单按日期查询}
\usage{
srcbilldatequery(token, category, fdate)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{应收单}

\item{fnumber}{2023-06-05}
}
\value{
返回值
}
\description{
源单按日期查询
}
\examples{
srcbilldatequery("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","销售出库单","2023-06-05")
}
