% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{srcbillnoquery}
\alias{srcbillnoquery}
\title{源单查询}
\usage{
srcbillnoquery(token, category, fnumber)
}
\arguments{
\item{token}{9B6F803F-9D37-41A2-BDA0-70A7179AF0F3}

\item{category}{销售出库单}

\item{fnumber}{D202305310016}
}
\value{
返回值
}
\description{
源单查询
}
\examples{
srcbillnoquery("9B6F803F-9D37-41A2-BDA0-70A7179AF0F3","销售出库单","D202305310016")
}
